Split comparison detail page helpers
